post-25.06: Update default branch  (#114)
* ci: Fix shape tensor segfault

* ci: Update SetDataFromShapeValues implementation after changing shape_values type (#113)

---------

+++ src/shape_tensor.h
@@ -1,4 +1,4 @@
-// Copyright 2024, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// Copyright 2024-2025, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions
@@ -51,7 +51,7 @@ class ShapeTensor {
       const char* input_name, bool support_batching, size_t total_batch_size);
 
   TRITONSERVER_Error* SetDataFromShapeValues(
-      const int32_t* shape_values, TRITONSERVER_DataType datatype,
+      const int64_t* shape_values, TRITONSERVER_DataType datatype,
       size_t nb_shape_values);
 
   int64_t GetDistance(const ShapeTensor& other, int64_t total_batch_size) const;
